+++ core/src/it/units/sdm/project/game/gui/GameScreen.java
@@ -3,18 +3,15 @@ package it.units.sdm.project.game.gui;
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Screen;
 import com.badlogic.gdx.graphics.Color;
-import com.badlogic.gdx.graphics.Pixmap;
-import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.graphics.g2d.TextureAtlas;
-import com.badlogic.gdx.graphics.g2d.TextureRegion;
 import com.badlogic.gdx.scenes.scene2d.*;
-import com.badlogic.gdx.scenes.scene2d.ui.Stack;
 import com.badlogic.gdx.scenes.scene2d.ui.*;
 import com.badlogic.gdx.scenes.scene2d.utils.Drawable;
 import com.badlogic.gdx.utils.Align;
 import com.badlogic.gdx.utils.ScreenUtils;
 import com.badlogic.gdx.utils.viewport.FitViewport;
+import it.units.sdm.project.board.gui.GuiBoard;
 import org.jetbrains.annotations.NotNull;
 
 import static it.units.sdm.project.game.gui.FreedomGame.NUMBER_OF_COLUMNS;
@@ -22,38 +19,26 @@ import static it.units.sdm.project.game.gui.FreedomGame.NUMBER_OF_ROWS;
 
 public class GameScreen implements Screen {
 
-
     public static final int TILE_SIZE = 75;
     public static final Color DARK_TILE = new Color(181 / 255f, 136 / 255f, 99 / 255f, 1);
     public static final Color LIGHT_TILE = new Color(240 / 255f, 217 / 255f, 181 / 255f, 1);
-    public static final Color HIGHLIGHT_DARK_TILE = new Color(105 / 255f, 105 / 255f, 105 / 255f, 255 / 255f);
-    public static final Color HIGHLIGHT_LIGHT_TILE = new Color(169 / 255f, 169 / 255f, 169 / 255f, 255 / 255f);
-    @NotNull
-    private final Table boardLayout;
     @NotNull
     private final FreedomGame game;
     @NotNull
     private final Stage stage;
     @NotNull
-    private Texture whiteSquareTexture;
-    @NotNull
     private final Skin skin;
     @NotNull
     private final Table container;
-    @NotNull
-    private final TextureAtlas atlas;
 
     public GameScreen(@NotNull FreedomGame game) {
         this.game = game;
         stage = new Stage(new FitViewport(1200, 640), new SpriteBatch());
-        boardLayout = new Table();
         container = new Table();
-        atlas = new TextureAtlas(Gdx.files.internal("uiskin.atlas"));
+        TextureAtlas atlas = new TextureAtlas(Gdx.files.internal("uiskin.atlas"));
         skin = new Skin(Gdx.files.internal("uiskin.json"));
-        @NotNull TextArea firstTextArea = new TextArea("Welcome to Freedom! Tap anywhere on the board to begin!\n", skin);
-        // init tile textures //
-        initTextures();
-        // the above part may be incorporated in a custom texture pack //
+        GuiBoard boardLayout = new GuiBoard(this, NUMBER_OF_ROWS, NUMBER_OF_COLUMNS, DARK_TILE, LIGHT_TILE);
+        TextArea firstTextArea = new TextArea("Welcome to Freedom! Tap anywhere on the board to begin!\n", skin);
         skin.addRegions(atlas);
         stage.addActor(container);
         Gdx.input.setInputProcessor(stage);
@@ -64,45 +49,6 @@ public class GameScreen implements Screen {
         firstTextArea.setDisabled(true);
         container.add(firstTextArea).expand().fill();
         container.add(boardLayout).width(NUMBER_OF_COLUMNS * TILE_SIZE);
-        initBoard();
-    }
-
-    private void initTextures() {
-        Pixmap whiteSquare = new Pixmap(TILE_SIZE, TILE_SIZE, Pixmap.Format.RGB565);
-        whiteSquare.setColor(Color.WHITE);
-        whiteSquare.fillRectangle(0, 0, TILE_SIZE, TILE_SIZE);
-        whiteSquareTexture = new Texture(whiteSquare);
-        whiteSquare.dispose();
-    }
-
-    private void initBoard() {
-        for (int i = 0; i < NUMBER_OF_ROWS; i++) {
-            boardLayout.row();
-            if (isIndexEven(i)) {
-                initBoardColumns(DARK_TILE, LIGHT_TILE);
-            } else {
-                initBoardColumns(LIGHT_TILE, DARK_TILE);
-            }
-        }
-    }
-
-    private void initBoardColumns(Color oddTilesColor, Color evenTilesColor) {
-        for (int j = 0; j < NUMBER_OF_COLUMNS; j++) {
-            TextureRegion whiteTextureRegion = new TextureRegion(whiteSquareTexture, 0, 0, TILE_SIZE, TILE_SIZE);
-            Image tile = new Image(whiteTextureRegion);
-            if (isIndexEven(j)) {
-                tile.setColor(evenTilesColor);
-            } else {
-                tile.setColor(oddTilesColor);
-            }
-            Stack tileAndPiece = new Stack(tile);
-            tileAndPiece.addListener(new TileClickListener(game, skin, stage, container));
-            boardLayout.add(tileAndPiece).size(TILE_SIZE);
-        }
-    }
-
-    private static boolean isIndexEven(int i) {
-        return i % 2 == 0;
     }
 
     @Override
@@ -142,9 +88,23 @@ public class GameScreen implements Screen {
     public void dispose() {
         stage.dispose();
         stage.getBatch().dispose();
-        whiteSquareTexture.dispose();
         skin.dispose();
         // the skin disposes of the atlas
-        atlas.dispose();
+    }
+
+    public @NotNull FreedomGame getGame() {
+        return game;
+    }
+
+    public @NotNull Stage getStage() {
+        return stage;
+    }
+
+    public @NotNull Skin getSkin() {
+        return skin;
+    }
+
+    public @NotNull Table getContainer() {
+        return container;
     }
 }
